Added file via upload
This script uses nvidia-smi to check the gpu usage, then makes the device with the least usage the visible gpu. This should help assign gpus for single gpu codes on a nodes with multiple gpu devices. Currently this code only checks device usage once, when the code is called. This could lead to potential errors as device usage naturally goes up and down. In testing for me, its worked so far.